refactor: 잘못된 변수명 수정 및 import 경로 정리
