+++ src/components/common/Modal/index.ts
@@ -1,1 +1,2 @@
 export { default as Modal } from './Modal';
+export { default as LoginForm } from './LoginForm';
