+++ src/hooks/index.ts
@@ -2,3 +2,5 @@ export { default as useForm } from './useForm';
 export { default as useAuth } from './useAuth';
 export { default as useUser } from './useUser';
 export { default as useModal } from './useModal';
+export { default as usePassword } from './usePassword';
+export { default as useInfiniteScroll } from './useInfiniteScroll';
