+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,6 +1,5 @@
 import { Link } from 'react-router-dom';
 import Warning from './Warning';
-import { Button } from '~/components/common';
 
 const NotFoundPage = () => {
   return (
