+++ src/pages/ProfilePage/ProfilePage.tsx
@@ -2,7 +2,6 @@ import dayjs from 'dayjs';
 import ProfileHeader from './ProfileHeader';
 import { Image } from '~/components/common';
 import { useUser } from '~/hooks';
-import { getRelativeTime } from '~/utils';
 
 const PostCard = ({ post }: { post: any }) => {
   const currentTime = dayjs();
