+++ src/routes/router.tsx
@@ -13,7 +13,7 @@ import {
   PostListPage,
   PostPage,
   ProfilePage,
-  SignupPage
+  SignUpPage
 } from '~/pages';
 
 const router = createBrowserRouter([
@@ -31,7 +31,7 @@ const router = createBrowserRouter([
           { path: PATH.PROFILE, element: <ProfilePage /> }
         ]
       },
-      { path: PATH.SIGNUP, element: <SignupPage /> },
+      { path: PATH.SIGNUP, element: <SignUpPage /> },
       { path: PATH.CHANNELS, element: <ChannelListPage /> },
       { path: PATH.POST, element: <PostPage /> },
       { path: PATH.POSTS, element: <PostListPage /> },
