+++ src/services/index.ts
@@ -2,3 +2,4 @@ export * from './channelService';
 export * from './commentService';
 export * from './postService';
 export * from './userService';
+export * from './notificationService';
